[src] Move composers for AudioSet
